% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popcountr.R
\name{popcountr}
\alias{popcountr}
\title{Fast population count}
\usage{
popcountr(x)
}
\arguments{
\item{x}{A single non-negative integer vector which can contain NA_integer_.
Results for negative integers are undefined.}
}
\value{
Population counts of each x.
}
\description{
Considerably fast population count using C++ compiler extensions and CPU instructions.
}
\examples{
popcountr(5)
popcountr(2147483647)
popcountr(c(5, 19))
popcountr(c(15, NA, 31))
}
