% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{popcountr_}
\alias{popcountr_}
\title{Fast population count}
\usage{
popcountr_(x)
}
\arguments{
\item{x}{A single integer vector.}
}
\value{
Population counts of each non-negative integer in x and
  an unspecified integer for NA_integer_. Results for negative integers
  are undefined.
}
\description{
A Rcpp implementation of population count using C++ compiler extensions and CPU instructions.
}
